fixed todays entries error
